A site for users to find and share meal recipes
Users are be able to sign up for an account, sign in, and build recipes
with ingredients and instructions. A search feature will provide users
the ability to find recipes.  Users arebe able to favorite recipes, and
their profile page features a list of the recipes they have favorited.
The website features a REST API.